+++ src/components/add-trip-event-button.js
@@ -1,4 +1,4 @@
-const createAddEventButtonTemplate = () => {
+const createAddTripEventButtonTemplate = () => {
   return (
     `<button
       class="trip-main__event-add-btn  btn  btn--big  btn--yellow"
@@ -9,4 +9,4 @@ const createAddEventButtonTemplate = () => {
   );
 };
 
-export {createAddEventButtonTemplate};
+export {createAddTripEventButtonTemplate};
