--- src/components/days.js
@@ -1,7 +1,0 @@
-const createDaysTemplate = () => {
-  return (
-    `<ul class="trip-days"></ul>`
-  );
-};
-
-export {createDaysTemplate};
