+++ src/components/trip-days.js
@@ -1,0 +1,7 @@
+const createTripDaysTemplate = () => {
+  return (
+    `<ul class="trip-days"></ul>`
+  );
+};
+
+export {createTripDaysTemplate};
