+++ src/components/trip-event-editor-destination.js
@@ -1,4 +1,4 @@
-const createAddEventDestinationTemplate = (photos) => {
+const createTripEventEditorDestinationTemplate = (photos) => {
   return (
     `<section class="event__section  event__section--destination">
       <h3 class="event__section-title  event__section-title--destination">Destination</h3>
@@ -17,4 +17,4 @@ const createAddEventDestinationTemplate = (photos) => {
   );
 };
 
-export {createAddEventDestinationTemplate};
+export {createTripEventEditorDestinationTemplate};
