+++ src/components/trip-statistics.js
